+++ test/collective/fleet/hybrid_parallel_communicate_group.py
@@ -141,73 +141,8 @@ class TestHybridEPGroup:
         assert mp_group.ranks == [0, 1]
 
 
-class TestHybridCPGroup:
-    def __init__(self):
-        paddle.distributed.init_parallel_env()
-        group_names = [
-            "moe_sharding",
-            "sharding",
-            "pipe",
-            "sep",
-            "data",
-            "expert",
-            "context",
-            "model",
-        ]
-        dims = [1, 1, 1, 1, 1, 2, 2, 2]
-
-        self.hcg = tp.EPHybridCommunicateGroup(group_names, dims)
-
-    def test_all(self):
-        global_rank = paddle.distributed.get_rank()
-
-        dp_rank = self.hcg.get_data_parallel_rank()
-        assert dp_rank == 0
-        assert self.hcg.get_expert_parallel_world_size() == 2
-        assert self.hcg.get_moe_sharding_parallel_world_size() == 1
-        assert self.hcg.get_model_parallel_world_size() == 2
-        assert self.hcg.get_expert_parallel_rank() == global_rank
-        assert self.hcg.get_moe_sharding_parallel_rank() == 0
-        assert self.hcg.get_expert_parallel_group_src_rank() == 0
-        assert (
-            self.hcg.get_moe_sharding_parallel_group_src_rank() == global_rank
-        )
-
-        moe_sharding_group = self.hcg.get_moe_sharding_parallel_group()
-        ep_group = self.hcg.get_expert_parallel_group()
-        mp_group = self.hcg.get_model_parallel_group()
-        assert moe_sharding_group.ranks == [global_rank]
-        assert ep_group.ranks == [0, 1]
-        assert mp_group.ranks == [0, 1]
-
-        assert self.hcg.get_context_parallel_rank() == global_rank
-        assert self.hcg.get_context_parallel_world_size() == 2
-        cp_group = self.hcg.get_context_parallel_group()
-        assert cp_group.ranks == [0, 1]
-        assert self.hcg.get_context_parallel_group_src_rank() == 0
-        cp_sharding_group = self.get_cp_sharding_parallel_group()
-        assert cp_sharding_group.ranks == [global_rank]
-        assert self.hcg.get_cp_sharding_parallel_group_src_rank() == global_rank
-        cp_mp_group = self.hcg.get_cp_mp_parallel_group()
-        assert cp_mp_group.ranks == [0, 1]
-        assert self.hcg.get_cp_mp_parallel_group_src_rank() == 0
-        assert self.hcg.get_sharding_parallel_world_size() == 1
-        assert (
-            self.hcg.get_sharding_parallel_world_size(
-                with_context_parallel=True
-            )
-            == 1
-        )
-        assert self.hcg.get_sharding_parallel_rank() == 0
-        assert (
-            self.hcg.get_sharding_parallel_rank(with_context_parallel=True) == 0
-        )
-
-
 if __name__ == "__main__":
     gpt = TestNewGroupAPI()
     gpt.test_all()
     ep_test = TestHybridEPGroup()
     ep_test.test_all()
-    cp_test = TestHybridCPGroup()
-    cp_test.test_all()
